+++ .env
@@ -1,5 +1,3 @@
 SPOTIFY_CLIENT_ID=4f8a0448747a497e99591f5c8983f2d7
 SPOTIFY_CLIENT_SECRET=b7ba599280d64d2ab32cf9cc9cbec47a
 SPOTIFY_REDIRECT_URI=http://127.0.0.1:8080/callback
-APP_SECRET_KEY=goodgroup
-MONGO_CLIENT=mongodb+srv://test:b11094@friendify.plioijt.mongodb.net/?retryWrites=true&w=majority
