+++ static/css/styles.css
@@ -42,7 +42,6 @@ body {
 .special{
   color: #1faa50;
   font-size: 40px;
-  text-shadow: 2px 2px 4px #000;
 }
 
 .image-container{
@@ -236,8 +235,6 @@ nav.navbar{
 .item:nth-child(5):hover:before {
   height: 100%;
 }
-
-
 
 
 /* EVERYTHING .welcome_back IS REACT TEST WORK */
