+++ templates/index.html
@@ -12,7 +12,7 @@
         <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
 
     </head>
-    <body style="background-color: {{ background_color }};">
+    <body>
         <div class="container">
             <nav class="navbar">
                 <div class="logo">
@@ -39,7 +39,7 @@
           <div class="random-fact-container">
             {% if is_logged_in %}
                 {% if random_statistic %}
-                <p class = "random-fact" style="color: {{ text_color }};">Did you know?<br> {{ random_statistic }} <span class ="special"> {{ special_name }}</span></p>
+                <p class = "random-fact">Did you know?<br> {{ random_statistic }} <span class ="special"> {{ special_name }}</span></p>
                     {% if image_url %}
                     <div class = "image-container">
                         <img src="{{ image_url }}" alt="Image" class = "random-fact-image">
@@ -49,7 +49,7 @@
 
                 {% endif %}
             <div class = "look-top"> 
-                <footer style="color: {{ text_color }};">Check out the stats tab for more features!</footer>
+                <footer>Check out the <span class = "friend">stats</span> tab for more features!</footer>
             </div>
             {% else %}
         </div>
@@ -61,7 +61,7 @@
             <div class="spacer2"></div>
 
             <div class = "page-marker">
-                <footer style="color: {{ text_color }};">HOME</footer>
+                <footer>HOME</footer>
             </div>
  
         <!--     REACT TEST WORK
